【修改描述】解决主分支与开发分支冲突


Match-id-bc0b0d4d16b041722f1904cd8f8c55adceaa68c0

+++ src/sdk/mirror_client.cpp
@@ -1385,7 +1385,7 @@ BResult MirrorClient::ListRemote(uint16_t nid, ListRequest &req, std::unordered_
     }
 
     if (rsp.num != 0) {
-        size_t objSize = 0;
+        int objSize = 0;
         auto statInfo = reinterpret_cast<ObjStat *>(mr.address);
         for (uint32_t i = 0; i < rsp.num; i++) {
             objSize = objs.size();
